Add files via upload
DePaul's Algs14 package is needed to run this application, though I included a sample output in the zip folder as proof that this application can run properly. The first column on the left deals with the number of values we are inputting and the columns in the center and on the right deal with how long it takes to run the application; how much the runtime is improving.